// components/frontend/src/lib/server/grpc/generated/hackathon/page_service.ts
export type PageServiceDefinition = typeof PageServiceDefinition;
export const PageServiceDefinition = {
  name: "PageService",
  fullName: "hackathon.PageService",
  methods: {
    list: {
      name: "List",
      requestType: ListRequest as typeof ListRequest,
      requestStream: false,
      responseType: ListResponse as typeof ListResponse,
      responseStream: false,
      options: {},
    },
    get: {
      name: "Get",
      requestType: GetRequest as typeof GetRequest,
      requestStream: false,
      responseType: GetResponse as typeof GetResponse,
      responseStream: false,
      options: {},
    },
    create: {
      name: "Create",
      requestType: CreateRequest as typeof CreateRequest,
      requestStream: false,
      responseType: CreateResponse as typeof CreateResponse,
      responseStream: false,
      options: {},
    },
    edit: {
      name: "Edit",
      requestType: EditRequest as typeof EditRequest,
      requestStream: false,
      responseType: EditResponse as typeof EditResponse,
      responseStream: false,
      options: {},
    },
    delete: {
      name: "Delete",
      requestType: DeleteRequest as typeof DeleteRequest,
      requestStream: false,
      responseType: DeleteResponse as typeof DeleteResponse,
      responseStream: false,
      options: {},
    },
    /** Reordering methods - backend manages order uniqueness */
    moveUp: {
      name: "MoveUp",
      requestType: MoveUpRequest as typeof MoveUpRequest,
      requestStream: false,
      responseType: MoveUpResponse as typeof MoveUpResponse,
      responseStream: false,
      options: {},
    },
    moveDown: {
      name: "MoveDown",
      requestType: MoveDownRequest as typeof MoveDownRequest,
      requestStream: false,
      responseType: MoveDownResponse as typeof MoveDownResponse,
      responseStream: false,
      options: {},
    },
    setOrder: {
      name: "SetOrder",
      requestType: SetOrderRequest as typeof SetOrderRequest,
      requestStream: false,
      responseType: SetOrderResponse as typeof SetOrderResponse,
      responseStream: false,
      options: {},
    },
  },
} as const;
// ...
export interface PageServiceClient<CallOptionsExt = {}> {
  list(request: DeepPartial<ListRequest>, options?: CallOptions & CallOptionsExt): Promise<ListResponse>;
  get(request: DeepPartial<GetRequest>, options?: CallOptions & CallOptionsExt): Promise<GetResponse>;
  create(request: DeepPartial<CreateRequest>, options?: CallOptions & CallOptionsExt): Promise<CreateResponse>;
  edit(request: DeepPartial<EditRequest>, options?: CallOptions & CallOptionsExt): Promise<EditResponse>;
  delete(request: DeepPartial<DeleteRequest>, options?: CallOptions & CallOptionsExt): Promise<DeleteResponse>;
  /** Reordering methods - backend manages order uniqueness */
  moveUp(request: DeepPartial<MoveUpRequest>, options?: CallOptions & CallOptionsExt): Promise<MoveUpResponse>;
  moveDown(request: DeepPartial<MoveDownRequest>, options?: CallOptions & CallOptionsExt): Promise<MoveDownResponse>;
  setOrder(request: DeepPartial<SetOrderRequest>, options?: CallOptions & CallOptionsExt): Promise<SetOrderResponse>;
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;