// components/frontend/src/lib/server/grpc/generated/hackathon/phase_service.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v2.11.6
//   protoc               unknown
// source: hackathon/phase_service.proto

/* eslint-disable */
import type { CallContext, CallOptions } from "nice-grpc-common";
import { CreateRequest } from "./messages/phase_svc/create_request";
import { CreateResponse } from "./messages/phase_svc/create_response";
import { DeleteRequest } from "./messages/phase_svc/delete_request";
import { DeleteResponse } from "./messages/phase_svc/delete_response";
import { EditRequest } from "./messages/phase_svc/edit_request";
import { EditResponse } from "./messages/phase_svc/edit_response";
import { GetRequest } from "./messages/phase_svc/get_request";
import { GetResponse } from "./messages/phase_svc/get_response";
import { ListRequest } from "./messages/phase_svc/list_request";
import { ListResponse } from "./messages/phase_svc/list_response";

export const protobufPackage = "hackathon";

export type PhaseServiceDefinition = typeof PhaseServiceDefinition;
export const PhaseServiceDefinition = {
  name: "PhaseService",
  fullName: "hackathon.PhaseService",
  methods: {
    list: {
      name: "List",
      requestType: ListRequest as typeof ListRequest,
      requestStream: false,
      responseType: ListResponse as typeof ListResponse,
      responseStream: false,
      options: {},
    },
    get: {
      name: "Get",
      requestType: GetRequest as typeof GetRequest,
      requestStream: false,
      responseType: GetResponse as typeof GetResponse,
      responseStream: false,
      options: {},
    },
    create: {
      name: "Create",
      requestType: CreateRequest as typeof CreateRequest,
      requestStream: false,
      responseType: CreateResponse as typeof CreateResponse,
      responseStream: false,
      options: {},
    },
    edit: {
      name: "Edit",
      requestType: EditRequest as typeof EditRequest,
      requestStream: false,
      responseType: EditResponse as typeof EditResponse,
      responseStream: false,
      options: {},
    },
    delete: {
      name: "Delete",
      requestType: DeleteRequest as typeof DeleteRequest,
      requestStream: false,
      responseType: DeleteResponse as typeof DeleteResponse,
      responseStream: false,
      options: {},
    },
  },
} as const;

export interface PhaseServiceImplementation<CallContextExt = {}> {
  list(request: ListRequest, context: CallContext & CallContextExt): Promise<DeepPartial<ListResponse>>;
  get(request: GetRequest, context: CallContext & CallContextExt): Promise<DeepPartial<GetResponse>>;
  create(request: CreateRequest, context: CallContext & CallContextExt): Promise<DeepPartial<CreateResponse>>;
  edit(request: EditRequest, context: CallContext & CallContextExt): Promise<DeepPartial<EditResponse>>;
  delete(request: DeleteRequest, context: CallContext & CallContextExt): Promise<DeepPartial<DeleteResponse>>;
}

export interface PhaseServiceClient<CallOptionsExt = {}> {
  list(request: DeepPartial<ListRequest>, options?: CallOptions & CallOptionsExt): Promise<ListResponse>;
  get(request: DeepPartial<GetRequest>, options?: CallOptions & CallOptionsExt): Promise<GetResponse>;
  create(request: DeepPartial<CreateRequest>, options?: CallOptions & CallOptionsExt): Promise<CreateResponse>;
  edit(request: DeepPartial<EditRequest>, options?: CallOptions & CallOptionsExt): Promise<EditResponse>;
  delete(request: DeepPartial<DeleteRequest>, options?: CallOptions & CallOptionsExt): Promise<DeleteResponse>;
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;