// components/frontend/src/lib/server/grpc/generated/vote/vote_service.ts
export type VoteServiceDefinition = typeof VoteServiceDefinition;
export const VoteServiceDefinition = {
  name: "VoteService",
  fullName: "vote.VoteService",
  methods: {
    /** VoteCategory CRUD */
    listVoteCategories: {
      name: "ListVoteCategories",
      requestType: ListVoteCategoriesRequest as typeof ListVoteCategoriesRequest,
      requestStream: false,
      responseType: ListVoteCategoriesResponse as typeof ListVoteCategoriesResponse,
      responseStream: false,
      options: {},
    },
    getVoteCategory: {
      name: "GetVoteCategory",
      requestType: GetVoteCategoryRequest as typeof GetVoteCategoryRequest,
      requestStream: false,
      responseType: GetVoteCategoryResponse as typeof GetVoteCategoryResponse,
      responseStream: false,
      options: {},
    },
    createVoteCategory: {
      name: "CreateVoteCategory",
      requestType: CreateVoteCategoryRequest as typeof CreateVoteCategoryRequest,
      requestStream: false,
      responseType: CreateVoteCategoryResponse as typeof CreateVoteCategoryResponse,
      responseStream: false,
      options: {},
    },
    editVoteCategory: {
      name: "EditVoteCategory",
      requestType: EditVoteCategoryRequest as typeof EditVoteCategoryRequest,
      requestStream: false,
      responseType: EditVoteCategoryResponse as typeof EditVoteCategoryResponse,
      responseStream: false,
      options: {},
    },
    deleteVoteCategory: {
      name: "DeleteVoteCategory",
      requestType: DeleteVoteCategoryRequest as typeof DeleteVoteCategoryRequest,
      requestStream: false,
      responseType: DeleteVoteCategoryResponse as typeof DeleteVoteCategoryResponse,
      responseStream: false,
      options: {},
    },
    /** Voting */
    submitVote: {
      name: "SubmitVote",
      requestType: SubmitVoteRequest as typeof SubmitVoteRequest,
      requestStream: false,
      responseType: SubmitVoteResponse as typeof SubmitVoteResponse,
      responseStream: false,
      options: {},
    },
    getVote: {
      name: "GetVote",
      requestType: GetVoteRequest as typeof GetVoteRequest,
      requestStream: false,
      responseType: GetVoteResponse as typeof GetVoteResponse,
      responseStream: false,
      options: {},
    },
    listVotes: {
      name: "ListVotes",
      requestType: ListVotesRequest as typeof ListVotesRequest,
      requestStream: false,
      responseType: ListVotesResponse as typeof ListVotesResponse,
      responseStream: false,
      options: {},
    },
    exportVotes: {
      name: "ExportVotes",
      requestType: ExportVotesRequest as typeof ExportVotesRequest,
      requestStream: false,
      responseType: ExportVotesResponse as typeof ExportVotesResponse,
      responseStream: false,
      options: {},
    },
    /** Vote Results */
    listVoteResults: {
      name: "ListVoteResults",
      requestType: ListVoteResultsRequest as typeof ListVoteResultsRequest,
      requestStream: false,
      responseType: ListVoteResultsResponse as typeof ListVoteResultsResponse,
      responseStream: false,
      options: {},
    },
    createVoteResult: {
      name: "CreateVoteResult",
      requestType: CreateVoteResultRequest as typeof CreateVoteResultRequest,
      requestStream: false,
      responseType: CreateVoteResultResponse as typeof CreateVoteResultResponse,
      responseStream: false,
      options: {},
    },
    editVoteResult: {
      name: "EditVoteResult",
      requestType: EditVoteResultRequest as typeof EditVoteResultRequest,
      requestStream: false,
      responseType: EditVoteResultResponse as typeof EditVoteResultResponse,
      responseStream: false,
      options: {},
    },
    deleteVoteResult: {
      name: "DeleteVoteResult",
      requestType: DeleteVoteResultRequest as typeof DeleteVoteResultRequest,
      requestStream: false,
      responseType: DeleteVoteResultResponse as typeof DeleteVoteResultResponse,
      responseStream: false,
      options: {},
    },
    suggestResults: {
      name: "SuggestResults",
      requestType: SuggestResultsRequest as typeof SuggestResultsRequest,
      requestStream: false,
      responseType: SuggestResultsResponse as typeof SuggestResultsResponse,
      responseStream: false,
      options: {},
    },
    exportResults: {
      name: "ExportResults",
      requestType: ExportResultsRequest as typeof ExportResultsRequest,
      requestStream: false,
      responseType: ExportResultsResponse as typeof ExportResultsResponse,
      responseStream: false,
      options: {},
    },
  },
} as const;
// ...
export interface VoteServiceClient<CallOptionsExt = {}> {
  /** VoteCategory CRUD */
  listVoteCategories(
    request: DeepPartial<ListVoteCategoriesRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ListVoteCategoriesResponse>;
  getVoteCategory(
    request: DeepPartial<GetVoteCategoryRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<GetVoteCategoryResponse>;
  createVoteCategory(
    request: DeepPartial<CreateVoteCategoryRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<CreateVoteCategoryResponse>;
  editVoteCategory(
    request: DeepPartial<EditVoteCategoryRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<EditVoteCategoryResponse>;
  deleteVoteCategory(
    request: DeepPartial<DeleteVoteCategoryRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<DeleteVoteCategoryResponse>;
  /** Voting */
  submitVote(
    request: DeepPartial<SubmitVoteRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<SubmitVoteResponse>;
  getVote(request: DeepPartial<GetVoteRequest>, options?: CallOptions & CallOptionsExt): Promise<GetVoteResponse>;
  listVotes(request: DeepPartial<ListVotesRequest>, options?: CallOptions & CallOptionsExt): Promise<ListVotesResponse>;
  exportVotes(
    request: DeepPartial<ExportVotesRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ExportVotesResponse>;
  /** Vote Results */
  listVoteResults(
    request: DeepPartial<ListVoteResultsRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ListVoteResultsResponse>;
  createVoteResult(
    request: DeepPartial<CreateVoteResultRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<CreateVoteResultResponse>;
  editVoteResult(
    request: DeepPartial<EditVoteResultRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<EditVoteResultResponse>;
  deleteVoteResult(
    request: DeepPartial<DeleteVoteResultRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<DeleteVoteResultResponse>;
  suggestResults(
    request: DeepPartial<SuggestResultsRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<SuggestResultsResponse>;
  exportResults(
    request: DeepPartial<ExportResultsRequest>,
    options?: CallOptions & CallOptionsExt,
  ): Promise<ExportResultsResponse>;
}

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;